Lesson 1.Comment fix

+++ Lesson_1/triangle_square.rb
@@ -1,7 +1,7 @@
 puts "Введите основание"
-a = gets.to_i
+a = gets.to_f
 puts "Введите высоту"
-h = gets.to_i
+h = gets.to_f
 
 square = a * h / 2
